Completely reshaped PositionEstimation class. First attempt, strange output. Probably led positions are not in the right measurement unit
Former-commit-id: ce505c623973b808a9750176f04ac3233893c0dd

+++ AutoportLinux/src/GenPurpFunc.hpp
@@ -142,7 +142,7 @@ namespace GenPurpFunc {
 			str += "\nPoint " + to_string(i+1) + ": [" + to_string(vect.at(i).pt.x) + ", " + to_string(vect.at(i).pt.y) + "]";
 		return str;
 	}
-	inline string pointVectorToString(const vector<Point3d> &vect) {
+	inline string pointVectorToString(const vector<Point3f> &vect) {
 		string str = "";
 		uint size = vect.size();
 		for(uint i = 0; i < size; i++)
